+++ lib-database-entities/src/commonMain/kotlin/com/ustadmobile/lib/db/entities/DownloadJobItemParentChildJoin.kt
@@ -1,9 +1,9 @@
 package com.ustadmobile.lib.db.entities
 
-import com.ustadmobile.lib.database.annotation.UmEntity
-import com.ustadmobile.lib.database.annotation.UmPrimaryKey
 import android.arch.persistence.room.Entity
 import android.arch.persistence.room.PrimaryKey
+import com.ustadmobile.lib.database.annotation.UmEntity
+import com.ustadmobile.lib.database.annotation.UmPrimaryKey
 
 @UmEntity
 @Entity
@@ -11,11 +11,11 @@ class DownloadJobItemParentChildJoin {
 
     @UmPrimaryKey(autoIncrement = true)
     @PrimaryKey(autoGenerate = true)
-    var djiPcjUid: Long = 0
+    var djiPcjUid: Int = 0
 
-    var djiParentDjiUid: Long = 0
+    var djiParentDjiUid: Int = 0
 
-    var djiChildDjiUid: Long = 0
+    var djiChildDjiUid: Int = 0
 
     var djiCepcjUid: Long = 0
 
@@ -28,7 +28,7 @@ class DownloadJobItemParentChildJoin {
         this.djiCepcjUid = cepcjUid
     }
 
-    constructor(djiParentDjiUid: Long, djiChildDjiUid: Long, djiCepcjUid: Long) {
+    constructor(djiParentDjiUid: Int, djiChildDjiUid: Int, djiCepcjUid: Long) {
         this.djiParentDjiUid = djiParentDjiUid
         this.djiChildDjiUid = djiChildDjiUid
         this.djiCepcjUid = djiCepcjUid
